port to 0.2 branch of ringo-sqlstore
I tried to change as little as possible but still ended up modifying
the model quiet a lot:

  * PageIndex is gone
  
I don't understand why this was here in the first place? Optimization for filebased DB? Instead of this, a page is now looked up via unique identifier `Page.slug`.
 
  *  The `name` and `body` properties of a Page instance default to the newest revision; see `Page.getByName(name, [revision])` and `page.revision` property

  * Revision entity

`name` is now versioned too